refactor: change structure of component Main
